requirements.txt
backports.shutil-get-terminal-size==1.0.0
boto==2.41.0
decorator==4.0.10
dj-database-url==0.4.1
Django==2.0.5
django-storages==1.4.1
gunicorn==19.6.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
simplegeneric==0.8.1
whitenoise==3.2
psycopg2==2.6.2
pylibmc==1.5.1
django-pylibmc==0.6.1
traitlets==4.2.1
djangorestframework==3.8.2

# ipdb==0.10.1
# ipython==6.1.0
# ipython-genutils==0.1.0
# locustio==0.8.1
